refactor(ui): extract toast and zoom notices from main.ts
Move the three pure DOM helpers (showToast, showZoomMessage,
hideZoomMessage) verbatim into src/ui/notices.ts. They depend only on
the DOM and their arguments, so they are the cleanest leaf to lift out
of the main.ts god module. No behavior change.

Verified: typecheck clean, production build OK, e2e 6 pass / 1 skip.
